refactor: move iconbutton to folder and update import location in components
