--- src/components/IconButton.jsx
@@ -1,10 +1,0 @@
-import { Button } from './ui/button';
-
-const IconButton = ({ icon, title, variant, onClick }) => (
-  <Button onClick={onClick} variant={variant} className="w-full justify-start">
-    {icon}
-    {title}
-  </Button>
-);
-
-export default IconButton;
